apps/extractor-documentos-docling/requirements.txt
requests>=2.31.0
pathlib
python-dotenv>=1.0.0